requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.10
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.16.0
snowballstemmer==2.2.0
tomli==2.2.1
tzdata==2024.1
